# Download page: Improved performance for chromium based browsers