I have this issue that when I selected all languages on the [download page](https://prismjs.com/download.html#themes=prism), Chrome freezes for about 12 seconds (!!) while using FireFox/Edge this operation takes an instant.

Btw. CPU: Intel i7-8700K @ 4.2GHz. (On my laptop, Chrome takes about 30sec)

This PR fixes this issue and after the fix, Chome takes less than a 1sec (not ideal but ok). FF and Edge stay fast.

<details>
<summary>
Here are some profiling results (Chrome):
</summary>

### Before

![image](https://user-images.githubusercontent.com/20878432/58354714-8f79b380-7e72-11e9-82e2-c196d832ac15.png)

TLDR;
Scripting: 177ms
Rendering: 11,485ms

### After

![image](https://user-images.githubusercontent.com/20878432/58355594-28113300-7e75-11e9-9360-04a135e4d01d.png)

TLDR;
Scripting: 121ms _// no idea why this is suddenly faster_
Rendering: 592ms

</details>